refactor(user-management): move user functions to accounnts context
