+++ apps/re/lib/accounts/accounts.ex
@@ -13,6 +13,19 @@ defmodule Re.Accounts do
 
   def query(query, _args), do: query
 
+  def change_role(user, new_role) do
+    user
+    |> User.update_changeset(%{role: new_role})
+    |> Repo.update()
+  end
+
+  def get_by_uuid(uuid) do
+    case Repo.get_by(User, uuid: uuid) do
+      nil -> {:error, :not_found}
+      user -> {:ok, user}
+    end
+  end
+
   def promote_user_to_admin(phone) do
     case Users.get_by_phone(phone) do
       {:ok, user} ->
